requirements.txt
aiogram==3.18.0
python-dotenv